Fix: Separate MyPageInfo components
