+++ src/components/MyPage/MyPageInfo/MyPageInfo.module.scss
@@ -163,78 +163,99 @@
   margin-top: 15px;
 }
 
-.nickname_input {
-  font-size: 14px;
+.input_wrapper {
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  max-width: 300px;
+  width: 100%;
+  height: 30px;
+  border: 1px solid #e5e5e5;
+  border-radius: 4px;
+  // justify-content: space-between;
+  // align-items: center;
+  .input {
+    height: 50px;
+    padding: 0 10px;
+    border: none;
+    // flex: 0 1 auto;
+    background-color: #fff;
+    font-size: 15px;
+    font-family: '-apple-system', HelveticaNeue, Roboto,
+      'Noto Sans KR', helvetica, Gulim, sans-serif;
+    color: #000;
+    box-shadow: 0 0 0 30px #fff inset !important;
+    transition: border 0.2s ease-in-out;
+    appearance: none;
+    box-sizing: border-box;
+    vertical-align: middle;
+    outline: 0;
+    margin: 0 1px;
+  }
+  .input_clear_button {
+    display: flex;
+    margin-right: 7px;
+    padding: 5px;
+    flex: 0 0 auto;
+    appearance: none;
+    border: none;
+    background: 0 0;
+    cursor: pointer;
+  }
+  .show_password_button {
+    display: flex;
+    background: url(https://static.msscdn.net/ui/build/m/img/login/ic-30-show-button.svg?v=20230119143941)
+      no-repeat 50% 50%;
+    margin-right: 7px;
+    padding: 0;
+    width: 30px;
+    height: 30px;
+    flex: 0 0 auto;
+    border: none;
+    cursor: pointer;
+  }
+  .hide_password_button {
+    display: flex;
+    background: url(https://static.msscdn.net/ui/build/m/img/login/ic-30-hide-button.svg?v=20230119143941)
+      no-repeat 50% 50%;
+    margin-right: 7px;
+    padding: 0;
+    width: 30px;
+    height: 30px;
+    flex: 0 0 auto;
+    border: none;
+    cursor: pointer;
+  }
+}
+
+.helper_container {
+  height: 30px;
+  .helper {
+    font-size: 12px;
+    margin-top: 1px;
+    margin-left: 2px;
+  }
+  .valid {
+    color: #14aaff;
+  }
+  .invalid {
+    color: red;
+  }
 }
 
 .nickname_constraint {
   padding-left: 18px;
-  color:#777777
+  color:#777777;
+  margin: 0;
 }
 
-.helper {
-  font-size: 12px;
-  margin-top: 1px;
-  margin-left: 2px;
-}
-.valid {
-  color: #14aaff;
-}
-.invalid {
-  color: red;
-}
-
-.label_checked {
-  min-width: 50px;
-  height: 30px;
-  padding: 6px 0;
-  float: left;
-  margin-right: -1px;
-  border: 1px solid #0078ff;
-  border-radius: 4px 0 0 4px;
-  font-size: 12px;
-  box-sizing: border-box;
-  color: white;
-  background: #0078ff;
-  text-align: center;
-  cursor: pointer;
-  z-index: 1;
-}
-
-.label_unchecked {
-  min-width: 50px;
-  height: 30px;
-  padding: 6px 0;
-  float: left;
-  margin-right: -1px;
-  border: 1px solid #ebebeb;
-  border-radius: 0 4px 4px 0;
-  font-size: 12px;
-  box-sizing: border-box;
-  color: #a5a5a5;
-  text-align: center;
-  cursor: pointer;
-  z-index: 1;
-}
-
-.edit_button {
-  position: absolute;
-  right: 0;
-  padding: 0 8px;
-  margin-left: 15px;
-  border: 1px solid #e5e5e5;
-  background-color: white;
-  min-width: 60px;
-  height: 32px;
-  line-height: 30px;
-  font-family: 'Apple SD Gothic Neo', 'Noto Sans KR', sans-serif !important;
-  box-sizing: border-box;
-  cursor: pointer;
-  transition: border 0.2s ease-in-out;
-}
-
-.edit_button:hover {
-  border-color: black;
+.password_label_input_wrapper {
+  display: flex;
+  margin-bottom: 20px;
+  .password_label {
+    min-width: 165px;
+    font-size: 14px;
+  }
 }
 
 .description_content {
